+++ Dialog.txt
@@ -1,13 +1,13 @@
 환자대사(8)
 1. 엑스레이
-	1) 아들이 장난감을 삼켰어요. 숨을 못쉬고 있어요. 제발 도와주세요.
-	2) 제 몸에 기생충이 있는것 같아요. 위를 갉아먹는 것 같아요. 기생충 좀 빼주세요.
+	1) 지원: 아들이 장난감을 삼켰어요. 숨을 못쉬고 있어요. 제발 도와주세요.
+	2) 재식: 제 몸에 기생충이 있는것 같아요. 위를 갉아먹는 것 같아요. 기생충 좀 빼주세요.
 	3) 저.. 몸.. 귀신.. 도와주세
 
 2. 청진기
-	1) 너무 많이 뛰었더니, 심장이 빨라져서 돌아오지 않아요. 도와주세요.
-	2) 제가 아끼는 여자친구와 헤어졌어요. 너무 슬퍼서 힘들어요. 몸에 무슨 문제가 있는지 확인해주세요.
-	3) 평소에 커피와 담배를 가까이 하는 직장인이에요. 요즘들어 심장이 빨리뛰는 것 같아요. 제게 문제가 있는걸까요?
+	1) 도현: 너무 많이 뛰었더니, 심장이 빨라져서 돌아오지 않아요. 도와주세요.
+	2) 지민: 제가 아끼는 여자친구와 헤어졌어요. 너무 슬퍼서 힘들어요. 몸에 무슨 문제가 있는지 확인해주세요.
+	3) 지은: 평소에 커피와 담배를 가까이 하는 직장인이에요. 요즘들어 심장이 빨리뛰는 것 같아요. 제게 문제가 있는걸까요?
 3. 주사기(링거)
 	1) 감기에 걸려서 링거를 맞고있어요. 근데 어지럽고 이상한 느낌이 들어요. 아까 흘려듣기로는 A팩을 받아야한다는데, 잘 진행되고 있는건가요?
 	2) 사고로 피를 너무 많이 흘렸어요. 수혈팩을 맞아야하는데, 도와주세요.
