+++ readme.txt
@@ -9,3 +9,15 @@ https://www.ursinaengine.org
 
 저작권 무료 음원사이트(Lo-Fi)
 https://uppbeat.io/browse/music/lo-fi-beats
+
+기획
+프로그래밍
+모델링 및 디자인
+음향
+
+
+Background music by PlayOnLoop.com
+Icon made by Studio Poppeople from www.flaticon.com
+
+Special Thanks
+사랑하는 반려견 미루에게 이 게임을 바칩니다.
